Move linux code to separate file
