+++ src/modules/cpu/cpu_stats.cpp
@@ -9,4 +9,9 @@ std::chrono::nanoseconds cpu_stats_get_steal_time()
     return 0ns;
 }
 
+utilization_time get_thread_utilization_time()
+{
+    return utilization_time{};
+}
+
 } // namespace openperf::cpu::internal
