+++ src/modules/cpu/cpu_stats.hpp
@@ -5,6 +5,13 @@
 
 namespace openperf::cpu::internal {
 
+struct utilization_time {
+    std::chrono::nanoseconds user;
+    std::chrono::nanoseconds system;
+    std::chrono::nanoseconds steal;
+};
+
+utilization_time get_thread_utilization_time();
 std::chrono::nanoseconds cpu_stats_get_steal_time();
 
 } // namespace openperf::cpu::internal
